% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.est.R
\name{extract.est}
\alias{extract.est}
\alias{extract.est,matrix,character,missing,missing-method}
\alias{extract.est,glm,character,logical,missing-method}
\title{Extract estimates and confidence intervals from a glm}
\usage{
extract.est(model, coefficient, ci, ci.clustered)

\S4method{extract.est}{matrix,character,missing,missing}(model, coefficient, ci)

\S4method{extract.est}{glm,character,logical,missing}(model, coefficient, ci)
}
\arguments{
\item{model}{A fitted glm model}

\item{coefficient}{A coefficient to extract statistics for}

\item{ci}{Whether to extract 95\% confidence intervals}

\item{ci.clustered}{Whether to extract a 95\% clustered confidence interval}
}
\description{
Extract estimates and confidence intervals from a glm
}
