% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.est.R
\name{extract.est}
\alias{extract.est}
\title{Extract}
\usage{
extract.est(model, coefficient, ci, ci.clustered)
}
\arguments{
\item{model}{A fitted glm model}

\item{coefficient}{A coefficient to extract statistics for}

\item{ci}{Whether to extract 95\% confidence intervals}
}
\description{
Extract
}
