% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\docType{methods}
\name{R2.CoxSnell}
\alias{R2.CoxSnell}
\alias{R2.CoxSnell,glm-gaussian-method}
\title{Cox-Snell R2}
\usage{
R2.CoxSnell(object)

\S4method{R2.CoxSnell}{glm-gaussian}(object)
}
\arguments{
\item{object}{A R2Wrap object}
}
\description{
Cox-Snell R2
}
