% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\docType{methods}
\name{R2.likelihood}
\alias{R2.likelihood}
\alias{R2.likelihood,R2Wrap-method}
\title{Sometimes referred to as Cohen R2 (probably from a textbook)}
\usage{
R2.likelihood(object)

\S4method{R2.likelihood}{R2Wrap}(object)
}
\arguments{
\item{object}{A R2Wrap object}
}
\description{
Sometimes referred to as Cohen R2 (probably from a textbook)
}
