% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\docType{methods}
\name{R2.Nagelkerke}
\alias{R2.Nagelkerke}
\alias{R2.Nagelkerke,R2Wrap-method}
\title{Nagelkerke's R2}
\usage{
R2.Nagelkerke(object)

\S4method{R2.Nagelkerke}{R2Wrap}(object)
}
\arguments{
\item{object}{A R2Wrap object}
}
\description{
Nagelkerke's R2
}
