% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\docType{methods}
\name{R2.scaleLiability}
\alias{R2.scaleLiability}
\alias{R2.scaleLiability,R2Wrap,numeric,numeric,numeric-method}
\alias{R2.scaleLiability,glm-gaussian,numeric,missing,missing-method}
\alias{R2.scaleLiability,glm-gaussian,vector,missing,missing-method}
\alias{R2.scaleLiability,glm-binomial,vector,missing,missing-method}
\alias{R2.scaleLiability,glm-binomial,numeric,missing,missing-method}
\title{Liability scale R2 of a glm model}
\usage{
R2.scaleLiability(object, prevalence, R2, prevalence.sample)

\S4method{R2.scaleLiability}{R2Wrap,numeric,numeric,numeric}(object, prevalence, R2, prevalence.sample)

\S4method{R2.scaleLiability}{`glm-gaussian`,numeric,missing,missing}(object, prevalence, R2, prevalence.sample)

\S4method{R2.scaleLiability}{`glm-gaussian`,vector,missing,missing}(object, prevalence, R2, prevalence.sample)

\S4method{R2.scaleLiability}{`glm-binomial`,vector,missing,missing}(object, prevalence)

\S4method{R2.scaleLiability}{`glm-binomial`,numeric,missing,missing}(object, prevalence)
}
\arguments{
\item{object}{A R2Wrap object}

\item{prevalence}{Population prevalence or a vector of prevalence}

\item{R2}{R-squared on the observed scale of the model}

\item{prevalence.sample}{The sample prevalence of Y=1}
}
\value{
Liability scale R2 given provided prevalence
}
\description{
Calculate liability scale R2 given a population prevalence or a set of prevalences
according to Lee et.al (2012).
}
\examples{
# Fit a glm model
m <- glm(y~x, data=pop)

# Create the pgsModel object
modR2 <- pgsModel(m, cases=100, controls=200)

# Calculate the liability scale R2 for a population prevalence
scaleLiabilityR2(modR2, prevalence=0.1)

# ... or for a sequence of prevalences
scaleLiabilityR2(modR2, prevalence=seq(0.01, 0.2, 0.01))
}
\references{
Lee, S.H., Goddard, M.e., Wray, N.R. and Visscher, P.M. (2012), A Better Coefficient of Determination for Genetic Profile Analysis. Genet. Epidemiol., 36: 214-224. https://doi.org/10.1002/gepi.21614
}
