% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\docType{methods}
\name{R2.scaleObserved}
\alias{R2.scaleObserved}
\alias{R2.scaleObserved,glm-binomial-method}
\alias{R2.scaleObserved,glm-gaussian-method}
\title{Observed scale R2}
\usage{
R2.scaleObserved(object)

\S4method{R2.scaleObserved}{glm-binomial}(object)

\S4method{R2.scaleObserved}{glm-gaussian}(object)
}
\description{
Variance in predictions/Variance in Y Var(Ŷ)/Var(Y)
}
