% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\docType{class}
\name{R2Wrap-class}
\alias{R2Wrap-class}
\title{An S4 class to contain a fitted model object}
\description{
An S4 class to contain a fitted model object
}
\section{Slots}{

\describe{
\item{\code{model}}{The type of model used}

\item{\code{family}}{The family link function.}

\item{\code{validModels}}{The accepted type of models}

\item{\code{cases}}{The number of \code{y=1}}

\item{\code{controls}}{The number of \code{y=0}}

\item{\code{comparison}}{A formula to compare models with.}

\item{\code{N}}{The sample size}
}}

