% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcompact.R
\name{regCompact}
\alias{regCompact}
\title{Run a series of regressions.}
\usage{
regCompact(phenotype, exposure, covariates, binary, dta, clustering = F)
}
\arguments{
\item{phenotype}{Variable to analyze.}

\item{exposure}{Exposure variable}

\item{covariates}{Variables to adjust the exposure for, eg "COV1 + COV2"}

\item{binary}{Pass TRUE to runa binomial/logistic model.}

\item{dta}{A data.frame containing all necessary variables}

\item{clustering}{Pass the name of the clustering variable in dta to add cluster robust confidence intervals.}
}
\value{
\code{regCompact} returns a data.frame with the following statistics.\tabular{ll}{
   \code{phenotype} \tab Phenotype variable name \cr
   \code{exposure} \tab Exposure variable name \cr
   \code{beta} \tab Beta coefficient \cr
   \code{stderr} \tab Standard error of \code{beta} \cr
   \code{z} \tab Z-statistic of \code{beta} \cr
   \code{p} \tab P-value of \code{beta} \cr
   \code{l95} \tab Lower bound of 95\% confidence interval \cr
   \code{u95} \tab Upper bound of 95\% confidence interval \cr
   \code{l95r} \tab Robust version of \code{l95} if clustering is used \cr
   \code{u95r} \tab Robust version of \code{u95} if clustering is used \cr
   \code{r2} \tab R-squared, for a continuous exposure and outcome, this is the square of \code{beta} \cr
    \tab For a binary phenotype, this is the observed scale R-squared \cr
   \code{r2Nagelkerke} \tab Nagelkerke's R-squared (if requested) \cr
   \code{family} \tab Regression family, either gaussian or binomial \cr
   \code{n} \tab Number of observations \cr
   \code{pheno_na} \tab Missingness in phenotype \cr
   \code{exp_na} \tab Missingness in exposure \cr
}
}
\description{
Run a regression phenotype~exposure + covariate 1 + covariate 2 and extract
parameters in a tabular format.
For binary=F, both phenotype and exposure will be standardized to mean 0 and
unit variance. For binary=T, exposure will be standardized.
For clustered data (eg samples from families) the clustering argument can be
used to calculate cluster robust confidence intervals (see sandwhich::vcovCL)
}
