% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{stdVar}
\alias{stdVar}
\title{Standardize a variable}
\usage{
stdVar(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A standardized variable
}
\description{
Standardizes a numeric vector to mean 0 and variance 1. Note that missing
values are ignored.
}
