% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\docType{methods}
\name{R2.scaleLiability}
\alias{R2.scaleLiability}
\alias{R2.scaleLiability,R2Wrap,numeric,numeric,numeric-method}
\alias{R2.scaleLiability,glm-gaussian,numeric,missing,missing-method}
\alias{R2.scaleLiability,glm-binomial,numeric,missing,missing-method}
\title{Liability scale R2 of a glm model}
\usage{
R2.scaleLiability(object, prevalence, R2, prevalence.sample)

\S4method{R2.scaleLiability}{R2Wrap,numeric,numeric,numeric}(object, prevalence, R2, prevalence.sample)

\S4method{R2.scaleLiability}{glm-gaussian,numeric,missing,missing}(object, prevalence)

\S4method{R2.scaleLiability}{glm-binomial,numeric,missing,missing}(object, prevalence)
}
\arguments{
\item{object}{A R2Wrap object}

\item{prevalence}{Population prevalence or a vector of prevalence}

\item{R2}{R-squared on the observed scale of the model}

\item{prevalence.sample}{The sample prevalence of Y=1}
}
\value{
Liability scale R2 given provided prevalence
}
\description{
Calculate liability scale R2 given a population prevalence or a set of prevalences
according to Lee et.al (2012).
}
\details{
The liability scale R2 as described in Lee et al is particularly
appropriate for ascertained case-control studies in which cases have been sampled
from the general population based on a binary phenotype \code{y}. The motivation comes
from genetic epidemiology for studying genetic liability (eg, R2 of a polygenic score) of \code{y}.
Other pseudo R2's may be more biased in such situations as the sample prevalence differs from the population
prevalance (eg 1:1 matching). However, to my knowledge it is not known how this study
how this R2 performs when the phenotype differs from the phenotype cases were ascertained for.
}
\examples{
\dontrun{
# Fit a linear glm model
m <- glm(y~x, data=data)

# Create the R2Wrap object
modR2 <- R2Wrap(m, cases=100, controls=200)

# If the fit is instead a binomial model
m <- glm(y~x, data=pop, family=binomial)
# the arguments cases and controls can be ignored

modR2 <- R2Wrap(m)

# Calculate the liability scale R2 for a population prevalence
scaleLiabilityR2(modR2, prevalence=0.1)

# ... or for a sequence of prevalences
scaleLiabilityR2(modR2, prevalence=seq(0.01, 0.2, 0.01))
}
}
\references{
Lee, S.H., Goddard, M.e., Wray, N.R. and Visscher, P.M. (2012), A Better Coefficient of Determination for Genetic Profile Analysis. Genet. Epidemiol., 36: 214-224. https://doi.org/10.1002/gepi.21614
}
