% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2wrap.R
\name{R2Wrap}
\alias{R2Wrap}
\title{Create a model container from a fitted glm model}
\usage{
R2Wrap(mod, cases = NA_integer_, controls = NA_integer_)
}
\arguments{
\item{mod}{The fitted model object}
}
\description{
Bla bla
}
